Move validation message functions to schemata utilities module
