+++ imod/mf6/validation.py
@@ -2,11 +2,10 @@
 This module contains specific validation utilities for Modflow 6.
 """
 
-from typing import Optional, cast
+from typing import cast
 
 import numpy as np
 
-from imod.common.statusinfo import NestedStatusInfo, StatusInfo, StatusInfoBase
 from imod.schemata import DimsSchema, NoDataComparisonSchema, ValidationError
 from imod.typing import GridDataArray
 
@@ -69,26 +68,3 @@ class DisBottomSchema(NoDataComparisonSchema):
             )
             if (overlaying_top_inactive & active).any():
                 raise ValidationError("inactive bottom above active cell")
-
-
-def validation_pkg_error_message(pkg_errors):
-    messages = []
-    for var, var_errors in pkg_errors.items():
-        messages.append(f"- {var}")
-        messages.extend(f"    - {error}" for error in var_errors)
-    return "\n" + "\n".join(messages)
-
-
-def pkg_errors_to_status_info(
-    pkg_name: str,
-    pkg_errors: dict[str, list[ValidationError]],
-    footer_text: Optional[str],
-) -> StatusInfoBase:
-    pkg_status_info = NestedStatusInfo(f"{pkg_name} package")
-    for var_name, var_errors in pkg_errors.items():
-        var_status_info = StatusInfo(var_name)
-        for var_error in var_errors:
-            var_status_info.add_error(str(var_error))
-        pkg_status_info.add(var_status_info)
-    pkg_status_info.set_footer_text(footer_text)
-    return pkg_status_info
